refactor(phase-2y): extract OpenProjectBulkCreateService
Continues openproject_client.py god-class decomposition (ADR-002).
The bulk-creation pipeline — three methods, the largest single
chunk remaining on the client — moves to
``OpenProjectBulkCreateService`` exposed as ``self.bulk_create``.

Methods moved
-------------
- bulk_create_records (generic mass-create for any Rails model;
  serializes records to JSON, transfers via docker, runs a
  self-contained Ruby script via rails runner or the persistent
  console, polls a result file with progress/stall detection,
  and returns a structured envelope. WorkPackage gets extra
  handling: Jira-key + provenance custom fields and an optional
  journal-creation template loaded from
  src/ruby/create_work_package_journals.rb)
- batch_create_work_packages (WP-specific public wrapper; thin
  entry point over performance_optimizer.batch_processor)
- _create_work_packages_batch (private WP-specific layer over
  bulk_create_records, called from OpenProjectWorkPackageService
  via self._client._create_work_packages_batch — the delegator on
  the client preserves that call shape)

Security analysis
-----------------
bulk_create_records — Ruby interpolations (all in script header):
- ``model_name = '{model}'`` — model is gated by
  ``client._validate_model_name`` (allowlist + ``^[A-Z][A-Za-z0-9]*$``
  regex). Safe.
- ``data_path = '{container_json.as_posix()}'`` — path is
  ``/tmp/{model.lower()}_bulk_{os.urandom(4).hex()}.json``;
  validated model + hex randomness only. Safe.
- ``result_path = '{container_result.as_posix()}'`` — path is
  ``/tmp/{result_name}`` where ``result_name`` is either a
  caller-supplied ``result_basename`` (only fed by internal
  callers using static strings or simple project-key
  interpolation) plus a hex/pid suffix, or
  ``bulk_result_{model.lower()}_{ts}_{pid}_{hex}.json``. Pre-
  existing behavior preserved — flagged here for completeness;
  no quote escaping is added in this mechanical move.
- ``ENV['J2O_BULK_PROGRESS_FILE'] ||= '{container_progress.as_posix()}'``
  — same family as result_path. Same verdict.
- ``# {_bulk_hint()}`` — sits inside a Ruby ``#`` comment (single
  line); content comes from controlled values
  (validated ``model``, ``config.jira_config['project_filter']``,
  ``datetime.now``, ``os.getpid()``). The ``_bulk_hint`` builder
  is wrapped in a try/except that returns a comment-safe fallback
  string — no newlines reach the comment line. Safe.
- Records themselves are never interpolated — they are written
  to a JSON file with ``json.dump`` and read back inside Ruby via
  ``JSON.parse(File.read(data_path))``. No injection surface.
- Runner script path ``f"/tmp/j2o_bulk_{os.urandom(4).hex()}.rb"``
  is hex-only; ``pkill -f`` and ``rails runner`` interpolations
  use that same hex path. Safe.
- No ``.to_json`` at the end of the script — the script writes
  ``JSON.generate(result)`` to a file via ``File.open(...) {...}``
  and the Python side reads the file back with ``json.load``,
  so no ``execute_query_to_json_file``/``.to_json`` double-encode
  hazard applies.
- No bare ``int()`` interpolation: numeric IDs in input ``records``
  ride through ``json.dump`` as JSON numbers and stay numeric
  inside Ruby.

batch_create_work_packages — no Ruby; returns immediately on
empty input and otherwise calls
``performance_optimizer.batch_processor.process_batches``. No
interpolation surface.

_create_work_packages_batch — Ruby interpolations:
- ``File.read('{container_json_path}')`` — the path is
  ``f"/tmp/j2o_batch_{batch_id}.json"`` with
  ``batch_id = uuid.uuid4().hex[:8]`` (hex chars only). Safe.
- All work-package fields are read inside Ruby from
  ``work_packages_data[...]`` after ``JSON.parse(File.read(...))``,
  so user-supplied subjects/descriptions/etc. never reach the
  Ruby source as text. Safe.
- The script's tail expression is a Ruby Hash literal
  (``{ created: ..., failed: ..., results: ... }``), and it's
  executed via ``execute_json_query`` which auto-appends
  ``.as_json`` (via the rails_runner_service). No explicit
  ``.to_json`` is appended, so no double-encode hazard. Safe.
- ``rm -f {container_json_path}`` cleanup uses the same hex-only
  path. Safe.

No security fixes were needed during the move — the original
implementations already used
``json.dump``+``File.read`` for record/data transfer and gated
the only Ruby-source-interpolated value (the model name)
through the ``_validate_model_name`` allowlist. The
``result_basename`` interpolation point is pre-existing behavior
and only fed by internal callers; flagged but not changed.

Side cleanups
-------------
- ``import subprocess`` and ``from datetime import UTC, datetime``
  are no longer used by ``openproject_client.py`` after the move;
  removed.

Numbers
-------
- openproject_client.py: 2342 -> 1622 LOC
- openproject_bulk_create_service.py: 0 -> 822 LOC

Verification
------------
- pytest tests/unit: 953 passed
- mypy src/: clean
- ruff check / ruff format: clean
